+++ app/src/main/java/com/example/studycompose/MainActivity.kt
@@ -1,50 +1,26 @@
 package com.example.studycompose
 
-import android.content.res.Configuration.UI_MODE_NIGHT_YES
+import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.spring
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.KeyboardArrowDown
-import androidx.compose.material.icons.filled.KeyboardArrowUp
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.studycompose.ui.theme.StudyComposeTheme
+import com.example.studycompose.ui.theme.ComposeAppTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            StudyComposeTheme {
+            ComposeAppTheme {
                 // A surface container using the 'background' color from the theme
                 MyApp(modifier = Modifier.fillMaxSize())
             }
@@ -52,14 +28,16 @@ class MainActivity : ComponentActivity() {
     }
 }
 
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 private fun MyApp(
     modifier: Modifier = Modifier,
 ) {
     var shouldShowOnboarding by rememberSaveable { mutableStateOf(true) }
 
-    Surface(
-        modifier
+    Scaffold(
+        modifier = modifier,
     ) {
         if (shouldShowOnboarding) {
             OnboardingScreen(onContinueClicked = { shouldShowOnboarding = false })
@@ -69,141 +47,10 @@ private fun MyApp(
     }
 }
 
-@Composable
-private fun Greetings(
-    modifier: Modifier = Modifier,
-    names: List<String> = List(1000) { "$it" }
-) {
-    LazyColumn(
-        modifier = Modifier.padding(vertical = 4.dp)
-    ) {
-        items(items = names) { name ->
-            Greeting(name = name)
-        }
-    }
-}
-
-@Composable
-private fun Greeting(name: String) {
-    Card(
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.primary
-        ),
-        modifier = Modifier.padding(vertical = 4.dp, horizontal = 8.dp)
-    ) {
-        CardContent(name = name)
-    }
-}
-
-@Composable
-private fun CardContent(name: String) {
-    var expanded by remember { mutableStateOf(false) }
-    val extraPadding by animateDpAsState(
-        if (expanded) 48.dp else 0.dp,
-        animationSpec = spring(
-            dampingRatio = Spring.DampingRatioMediumBouncy,
-            stiffness = Spring.StiffnessLow
-        )
-    )
-
-
-    Row(
-        modifier = Modifier
-            .padding(12.dp)
-            .animateContentSize(
-                animationSpec = spring(
-                    dampingRatio = Spring.DampingRatioMediumBouncy,
-                    stiffness = Spring.StiffnessLow
-                )
-            )
-    ) {
-        Column(
-            modifier = Modifier
-                .weight(1f)
-                .padding(12.dp)
-        ) {
-            Text(text = "Hello,")
-            Text(
-                text = "$name", style = MaterialTheme.typography.headlineMedium.copy(
-                    fontWeight = FontWeight.ExtraBold
-                )
-            )
-            if (expanded) {
-                Text(
-                    text = (stringResource(id = R.string.show_more_text) +
-                            stringResource(id = R.string.show_more_text2)).repeat(4),
-                )
-            }
-        }
-        IconButton(onClick = {
-            expanded = !expanded
-        }) {
-            Icon(
-                imageVector = if (expanded) Icons.Filled.KeyboardArrowUp else Icons.Filled.KeyboardArrowDown,
-                contentDescription = if (expanded) {
-                    stringResource(id = R.string.show_less)
-                } else {
-                    stringResource(id = R.string.show_more)
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun OnboardingScreen(
-    modifier: Modifier = Modifier,
-    onContinueClicked: () -> Unit
-) {
-    Column(
-        modifier = modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text("Welcome to Japan!")
-        Button(
-            modifier = Modifier.padding(vertical = 24.dp),
-            onClick = onContinueClicked
-        ) {
-            Text("Continue")
-        }
-    }
-}
-
-@Preview(
-    showBackground = true,
-    widthDp = 320,
-    uiMode = UI_MODE_NIGHT_YES,
-    name = "Dark Mode"
-)
-@Preview(showBackground = true, widthDp = 320)
-@Composable
-private fun DefaultPreview() {
-    StudyComposeTheme {
-        Greetings()
-    }
-}
-
 @Preview(showBackground = true, widthDp = 320)
 @Composable
 private fun MyAppPreview() {
-    StudyComposeTheme {
+    ComposeAppTheme {
         MyApp(modifier = Modifier.fillMaxSize())
-    }
-}
-
-@Preview(showBackground = true, widthDp = 320, heightDp = 320)
-@Composable
-private fun OnboardingPreview() {
-    StudyComposeTheme {
-        OnboardingScreen(onContinueClicked = {})
-    }
-}
-
-@Preview(showBackground = true, widthDp = 320)
-@Composable
-private fun GreetingsPreview() {
-    StudyComposeTheme {
-        Greetings()
     }
 }
